Flatten chains

Flatten a cia_chains object into a single cia_chain object. This is helpful for when you want to calculate a feature across using all samples across the cia_chains.

Usage

FlattenChains(chains)

Arguments

chains

A cia_chains object.

Examples

data <- bnlearn::learning.test

dag <- UniformlySampleDAG(colnames(data))
partitioned_nodes <- DAGtoPartition(dag)

scorer <- list(
  scorer = BNLearnScorer, 
  parameters = list(data = data)
  )

results <- SampleChains(10, partitioned_nodes, PartitionMCMC(), scorer)
FlattenChains(results)
#> $state
#> $state[[1]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         2    F
#> 4         3    D
#> 5         4    E
#> 6         5    B
#> 
#> $state[[2]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         2    F
#> 4         3    D
#> 5         4    E
#> 6         5    B
#> 
#> $state[[3]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         2    F
#> 4         3    D
#> 5         4    E
#> 6         5    B
#> 
#> $state[[4]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         2    F
#> 4         3    B
#> 5         4    E
#> 6         5    D
#> 
#> $state[[5]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         2    F
#> 4         3    B
#> 5         4    E
#> 6         5    D
#> 
#> $state[[6]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         2    F
#> 4         3    B
#> 5         4    E
#> 6         5    D
#> 
#> $state[[7]]
#>   partition node
#> 1         1    C
#> 2         2    F
#> 3         3    B
#> 4         4    E
#> 5         5    A
#> 6         6    D
#> 
#> $state[[8]]
#>   partition node
#> 1         1    C
#> 2         2    F
#> 3         3    B
#> 4         4    A
#> 5         4    E
#> 6         5    D
#> 
#> $state[[9]]
#>   partition node
#> 1         1    C
#> 2         2    F
#> 3         3    B
#> 4         4    A
#> 5         4    E
#> 6         5    D
#> 
#> $state[[10]]
#>   partition node
#> 1         1    C
#> 2         2    F
#> 3         3    B
#> 4         4    A
#> 5         4    E
#> 6         5    D
#> 
#> $state[[11]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    F
#> 4         2    D
#> 5         3    B
#> 6         3    E
#> 
#> $state[[12]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    F
#> 4         2    D
#> 5         3    B
#> 6         3    E
#> 
#> $state[[13]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    E
#> 4         2    D
#> 5         3    B
#> 6         3    F
#> 
#> $state[[14]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    E
#> 4         2    D
#> 5         3    B
#> 6         3    F
#> 
#> $state[[15]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    E
#> 4         2    D
#> 5         3    B
#> 6         3    F
#> 
#> $state[[16]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    E
#> 4         2    D
#> 5         3    B
#> 6         3    F
#> 
#> $state[[17]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    E
#> 4         2    D
#> 5         3    B
#> 6         3    F
#> 
#> $state[[18]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    E
#> 4         2    D
#> 5         3    B
#> 6         3    F
#> 
#> $state[[19]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         1    E
#> 4         2    F
#> 5         3    B
#> 6         3    D
#> 
#> $state[[20]]
#>   partition node
#> 1         1    A
#> 2         1    C
#> 3         2    E
#> 4         3    F
#> 5         4    B
#> 6         4    D
#> 
#> 
#> $proposal_info
#> $proposal_info[[1]]
#> $proposal_info[[1]]$proposal_used
#> [1] "split_join"
#> 
#> 
#> $proposal_info[[2]]
#> $proposal_info[[2]]$proposal_used
#> [1] "split_join"
#> 
#> 
#> $proposal_info[[3]]
#> $proposal_info[[3]]$proposal_used
#> [1] "split_join"
#> 
#> 
#> $proposal_info[[4]]
#> $proposal_info[[4]]$proposal_used
#> [1] "swap_node"
#> 
#> 
#> $proposal_info[[5]]
#> $proposal_info[[5]]$proposal_used
#> [1] "node_move"
#> 
#> 
#> $proposal_info[[6]]
#> $proposal_info[[6]]$proposal_used
#> [1] "split_join"
#> 
#> 
#> $proposal_info[[7]]
#> $proposal_info[[7]]$proposal_used
#> [1] "node_move"
#> 
#> 
#> $proposal_info[[8]]
#> $proposal_info[[8]]$proposal_used
#> [1] "node_move"
#> 
#> 
#> $proposal_info[[9]]
#> $proposal_info[[9]]$proposal_used
#> [1] "swap_adjacent"
#> 
#> 
#> $proposal_info[[10]]
#> $proposal_info[[10]]$proposal_used
#> [1] "node_move"
#> 
#> 
#> $proposal_info[[11]]
#> $proposal_info[[11]]$proposal_used
#> [1] "split_join"
#> 
#> 
#> $proposal_info[[12]]
#> $proposal_info[[12]]$proposal_used
#> [1] "split_join"
#> 
#> 
#> $proposal_info[[13]]
#> $proposal_info[[13]]$proposal_used
#> [1] "swap_node"
#> 
#> 
#> $proposal_info[[14]]
#> $proposal_info[[14]]$proposal_used
#> [1] "swap_node"
#> 
#> 
#> $proposal_info[[15]]
#> $proposal_info[[15]]$proposal_used
#> [1] "swap_node"
#> 
#> 
#> $proposal_info[[16]]
#> $proposal_info[[16]]$proposal_used
#> [1] "node_move"
#> 
#> 
#> $proposal_info[[17]]
#> $proposal_info[[17]]$proposal_used
#> [1] "stay_still"
#> 
#> 
#> $proposal_info[[18]]
#> $proposal_info[[18]]$proposal_used
#> [1] "node_move"
#> 
#> 
#> $proposal_info[[19]]
#> $proposal_info[[19]]$proposal_used
#> [1] "swap_node"
#> 
#> 
#> $proposal_info[[20]]
#> $proposal_info[[20]]$proposal_used
#> [1] "node_move"
#> 
#> 
#> 
#> $mcmc_info
#> $mcmc_info[[1]]
#> $mcmc_info[[1]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[1]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[1]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[1]]$jac
#> [1] 0.1541507
#> 
#> $mcmc_info[[1]]$mhr
#> [1] 420.1642
#> 
#> 
#> $mcmc_info[[2]]
#> $mcmc_info[[2]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[2]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[2]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[2]]$jac
#> [1] -0.1541507
#> 
#> $mcmc_info[[2]]$mhr
#> [1] -420.1642
#> 
#> 
#> $mcmc_info[[3]]
#> $mcmc_info[[3]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[3]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[3]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[3]]$jac
#> [1] 0.1823216
#> 
#> $mcmc_info[[3]]$mhr
#> [1] -16.91651
#> 
#> 
#> $mcmc_info[[4]]
#> $mcmc_info[[4]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[4]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[4]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[4]]$jac
#> [1] 0
#> 
#> $mcmc_info[[4]]$mhr
#> [1] 122.4647
#> 
#> 
#> $mcmc_info[[5]]
#> $mcmc_info[[5]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[5]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[5]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[5]]$jac
#> [1] 0.2231436
#> 
#> $mcmc_info[[5]]$mhr
#> [1] -555.965
#> 
#> 
#> $mcmc_info[[6]]
#> $mcmc_info[[6]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[6]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[6]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[6]]$jac
#> [1] 0.1823216
#> 
#> $mcmc_info[[6]]$mhr
#> [1] -17.25859
#> 
#> 
#> $mcmc_info[[7]]
#> $mcmc_info[[7]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[7]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[7]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[7]]$jac
#> [1] -0.1823216
#> 
#> $mcmc_info[[7]]$mhr
#> [1] 103.1993
#> 
#> 
#> $mcmc_info[[8]]
#> $mcmc_info[[8]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[8]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[8]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[8]]$jac
#> [1] 0.1823216
#> 
#> $mcmc_info[[8]]$mhr
#> [1] 47.31148
#> 
#> 
#> $mcmc_info[[9]]
#> $mcmc_info[[9]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[9]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[9]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[9]]$jac
#> [1] 0
#> 
#> $mcmc_info[[9]]$mhr
#> [1] -186.0552
#> 
#> 
#> $mcmc_info[[10]]
#> $mcmc_info[[10]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[10]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[10]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[10]]$jac
#> [1] -0.1823216
#> 
#> $mcmc_info[[10]]$mhr
#> [1] -91.72392
#> 
#> 
#> $mcmc_info[[11]]
#> $mcmc_info[[11]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[11]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[11]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[11]]$jac
#> [1] -0.3566749
#> 
#> $mcmc_info[[11]]$mhr
#> [1] 76.61861
#> 
#> 
#> $mcmc_info[[12]]
#> $mcmc_info[[12]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[12]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[12]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[12]]$jac
#> [1] -0.5306283
#> 
#> $mcmc_info[[12]]$mhr
#> [1] -1811.49
#> 
#> 
#> $mcmc_info[[13]]
#> $mcmc_info[[13]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[13]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[13]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[13]]$jac
#> [1] 0
#> 
#> $mcmc_info[[13]]$mhr
#> [1] 108.5444
#> 
#> 
#> $mcmc_info[[14]]
#> $mcmc_info[[14]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[14]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[14]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[14]]$jac
#> [1] 0
#> 
#> $mcmc_info[[14]]$mhr
#> [1] -108.5444
#> 
#> 
#> $mcmc_info[[15]]
#> $mcmc_info[[15]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[15]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[15]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[15]]$jac
#> [1] 0
#> 
#> $mcmc_info[[15]]$mhr
#> [1] -945.1964
#> 
#> 
#> $mcmc_info[[16]]
#> $mcmc_info[[16]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[16]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[16]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[16]]$jac
#> [1] 0
#> 
#> $mcmc_info[[16]]$mhr
#> [1] -426.181
#> 
#> 
#> $mcmc_info[[17]]
#> $mcmc_info[[17]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[17]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[17]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[17]]$jac
#> [1] 0
#> 
#> $mcmc_info[[17]]$mhr
#> [1] 0
#> 
#> 
#> $mcmc_info[[18]]
#> $mcmc_info[[18]]$accept
#> [1] FALSE
#> 
#> $mcmc_info[[18]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[18]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[18]]$jac
#> [1] 0
#> 
#> $mcmc_info[[18]]$mhr
#> [1] -426.181
#> 
#> 
#> $mcmc_info[[19]]
#> $mcmc_info[[19]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[19]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[19]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[19]]$jac
#> [1] 0
#> 
#> $mcmc_info[[19]]$mhr
#> [1] 112.0345
#> 
#> 
#> $mcmc_info[[20]]
#> $mcmc_info[[20]]$accept
#> [1] TRUE
#> 
#> $mcmc_info[[20]]$white_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[20]]$black_obeyed
#> [1] TRUE
#> 
#> $mcmc_info[[20]]$jac
#> [1] -0.2231436
#> 
#> $mcmc_info[[20]]$mhr
#> [1] 216.0217
#> 
#> 
#> 
#> $log_score
#>  [1] -24293.91 -24293.91 -24293.91 -24171.45 -24171.45 -24171.45 -24068.25
#>  [8] -24020.94 -24020.94 -24020.94 -24637.46 -24637.46 -24528.92 -24528.92
#> [15] -24528.92 -24528.92 -24528.92 -24528.92 -24416.88 -24200.86
#> 
#> attr(,"class")
#> [1] "cia_chain"